// server/src/types.ts
export interface TimeSeriesPoint {
    timestamp: string;
    value: number;
}

export interface TimeSeriesData {
    [key: string]: {
        unit: string;
        data: TimeSeriesPoint[];
    };
}

export interface AssetEvent {
    timestamp: string;
    description: string;
}

export interface Asset {
    id: string;
    name: string;
    siteId: string;
    siteName: string;
    type: 'Sensor' | 'Valve' | 'Controller' | 'Machine';
    health: number;
    criticality: 'Critical' | 'High' | 'Medium' | 'Low';
    status: string;
    namurStatus: 'F' | 'C' | 'S' | 'M' | 'OK';
    isSif: boolean;
    make: string;
    model: string;
    revision: string;
    serialNumber: string;
    eventTimestamp: string;
    calibrationDueDate: string;
    alert: {
        title: string;
        description: string;
        causeCode: string;
        remedyCode: string;
        remedies: { id: number; description: string; reference: string }[];
    };
    timeSeriesData: TimeSeriesData;
    events: AssetEvent[];
    history?: {day: string; status: 'F' | 'C' | 'S' | 'M' | 'OK'}[];
    cmmsStatus?: 'Pending' | 'InProgress' | 'Success' | 'Error';
    sapWoNumber?: string;
}

export interface Unit {
    id: string;
    name: string;
    assets: Asset[];
}

export interface Area {
    id:string;
    name: string;
    units: Unit[];
}

export interface WorkOrder {
  id: string; // SAP WO Number
  assetId: string;
  assetName: string;
  siteId: string;
  description: string;
  type: 'Work Order' | 'Work Request';
  status: 'Created' | 'Released' | 'In Progress' | 'Awaiting Parts' | 'Completed' | 'Closed' | 'Overdue';
  priority: 'High' | 'Medium' | 'Low';
  createdDate: string;
  dueDate: string;
}